+++ packages/base/index.js
@@ -5,7 +5,7 @@ module.exports = {
 
     // - Parseur
     parserOptions: {
-        ecmaVersion: 2020,
+        ecmaVersion: 2021,
         sourceType: 'script',
     },