+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1.0 (unreleased)
 - Met à jour les dépendances (et donc ajoute le support de Typescript 4.x).
+- Met à jour la version d'ecmascript supportée (=> 2021).
 - Met à jour la règle `@typescript-eslint/ban-ts-comment`: Un commentaire explicatif 
   est maintenant exigé lors de l'utilisation de `// @ts-expect-error` et `// @ts-ignore`.
 - Nouvelles règles: